Refactor the formatter to be much less icky
